generador_mvp/requirements.txt
pandas==1.5.3
google-cloud-pubsub==2.18.1
google-api-core==1.34.0
proto-plus==1.23.0
apache-beam==2.51.0
google-apitools==0.5.32
google-cloud-bigquery==3.13.0
protobuf==3.20.3